+++ pages/auth/signin.tsx
@@ -1,17 +1,9 @@
-import React from 'react';
-
 import {GetServerSideProps} from 'next';
-import {getProviders, getSession, signIn} from 'next-auth/client';
-import {Button} from 'react-bootstrap';
+import {getProviders, getSession} from 'next-auth/client';
 
-import {providerIcon} from '../../src/components/elements/common/userControl/icons';
+import {SignInPage, SignInPageProps} from '../../src/components/elements/common/userControl/signIn/main';
 import {GeneralPath} from '../../src/const/path/definitions';
-import {useI18n} from '../../src/i18n/hook';
 
-
-type SignInPageProps = {
-  providers: UnwrapPromise<ReturnType<typeof getProviders>>,
-}
 
 export const getServerSideProps: GetServerSideProps<SignInPageProps> = async (context) => {
   const {req, res, query} = context;
@@ -35,31 +27,6 @@ export const getServerSideProps: GetServerSideProps<SignInPageProps> = async (co
       providers: await getProviders(),
     },
   };
-};
-
-const SignInPage = ({providers}: SignInPageProps) => {
-  const {t} = useI18n();
-
-  if (!providers) {
-    return (
-      <h3 className="text-danger">
-        {t((t) => t.message.error.auth.noProvider)}
-      </h3>
-    );
-  }
-
-  return (
-    <div className="center-screen">
-      {Object.values(providers).map((provider) => (
-        <Button
-          key={provider.name} size="lg" variant="outline-light"
-          onClick={() => signIn(provider.id)}
-        >
-          {providerIcon[provider.name] || provider.name}
-        </Button>
-      ))}
-    </div>
-  );
 };
 
 export default SignInPage;
