--- src/components/elements/common/userControl/icons.tsx
@@ -1,6 +1,0 @@
-import React from 'react';
-
-
-export const providerIcon: Record<string, React.ReactNode> = {
-  Google: <i className="bi bi-google"/>,
-};
