+++ src/components/elements/common/userControl/signIn/icons.tsx
@@ -1,0 +1,12 @@
+import React from 'react';
+
+
+// Available Icons: https://icons.getbootstrap.com/#icons
+
+export const providerIcon: Record<string, React.ReactNode> = {
+  Google: <i className="bi bi-google"/>,
+  Discord: <i className="bi bi-discord"/>,
+  GitHub: <i className="bi bi-github"/>,
+  Twitch: <i className="bi bi-twitch"/>,
+  Twitter: <i className="bi bi-twitter"/>,
+};
