+++ src/components/elements/common/userControl/signIn/main.module.scss
@@ -1,0 +1,15 @@
+@use "../../../../../../public/colors";
+
+$button-bg-color: darken(colors.$color-black-50, 5%);
+$button-distance: 1rem;
+
+.providerButton {
+  border: 0;
+  background-color: $button-bg-color;
+  font-size: 2rem;
+  margin-left: $button-distance;
+
+  &:first-child {
+    margin-left: 0;
+  }
+}
